third_party/blink/web_tests/svg/dynamic-updates/SVGMarkerElement-svgdom-markerHeight-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'markerHeight' property of the SVGMarkerElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS markerElement.markerHeight.baseVal.value is 0
PASS markerElement.markerHeight.baseVal.value is 2
PASS successfullyParsed is true

TEST COMPLETE
